change move ICheckboxes to types
